Split off tests from big testbenches to auxiliary testbenches
Move resource copy tests and docker chown tests
Need to study the duplication of resources more closely
