+++ src/test/resources/testbash/auxiliary/check_bash_version.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+
+# VIASH START
+par_input="check_bash_version.sh"
+par_optional="help"
+resources_dir="."
+# VIASH END
+
+set -e
+
+bash --version
