+++ src/test/resources/testbash/auxiliary/config_resource_unsupported_protocol.vsh.yaml
