+++ src/test/resources/testbash/auxiliary/resource1.txt
@@ -1,0 +1,2 @@
+if you can read this,
+the resource was read correctly
