+++ src/test/resources/testbash/auxiliary/resource2.txt
@@ -1,0 +1,1 @@
+this file is only for testing
