+++ src/test/resources/testbash/auxiliary/resource3.txt
@@ -1,0 +1,1 @@
+yet another file that is only for testing
