+++ src/test/resources/testbash/auxiliary/resource_folder/resource_L1_2.txt
