+++ src/test/scala/com/dataintuitive/viash/MainBuildDockerSuite.scala
@@ -32,14 +32,6 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
   private val functionalityRequirements = Config.read(configRequirementsFile, modifyFun = false).functionality
   private val executableRequirementsFile = Paths.get(tempFolStr, functionalityRequirements.name).toFile
 
-  private val configResourcesCopyFile = getClass.getResource("/testbash/config_resource_test.vsh.yaml").getPath
-  private val configResourcesUnsupportedProtocolFile = getClass.getResource("/testbash/config_resource_unsupported_protocol.vsh.yaml").getPath
-
-  private val configDockerOptionsChownFile = getClass.getResource("/testbash/docker_options/config_chown.vsh.yaml").getPath
-  private val configDockerOptionsChownTwoOutputFile = getClass.getResource("/testbash/docker_options/config_chown_two_output.vsh.yaml").getPath
-  private val configDockerOptionsChownMultipleOutputFile = getClass.getResource("/testbash/docker_options/config_chown_multiple_output.vsh.yaml").getPath
-
-
   //<editor-fold desc="Test benches to build a generic script and run various commands to see if the functionality is correct">
   // convert testbash
   test("viash can create an executable") {
@@ -707,259 +699,6 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
     removeDockerImage("viash_requirement_apt")
   }
 
-  //</editor-fold>
-  //<editor-fold desc="Verify correct copying of resources">
-  test("Check resources are copied from and to the correct location") {
-
-    // copy some resources to /tmp/viash_tmp_resources/ so we can test absolute path resources
-    val tmpFolderResourceSourceFile = Paths.get(getClass.getResource("/testbash/resource3.txt").getFile)
-
-    val tmpFolderResourceDestinationFolder = Paths.get("/tmp/viash_tmp_resources/").toFile
-    val tmpFolderResourceDestinationFile = Paths.get(tmpFolderResourceDestinationFolder.getPath, "resource3.txt")
-
-    if (!tmpFolderResourceDestinationFolder.exists())
-      tmpFolderResourceDestinationFolder.mkdir()
-
-    Files.copy(tmpFolderResourceSourceFile, tmpFolderResourceDestinationFile, StandardCopyOption.REPLACE_EXISTING)
-
-    // generate viash script
-    TestHelper.testMain(
-      "build",
-      "-p", "docker",
-      "-o", tempFolStr,
-      configResourcesCopyFile
-    )
-
-    assert(executable.exists)
-    assert(executable.canExecute)
-
-    // List all expected resources and their md5sum
-    val expectedResources = List(
-      //("check_bash_version.sh", "0c3c134d4ff0ea3a4a3b32e09fb7c100"),
-      ("code.sh", "efa9e1aa1c5f2a0b91f558ead5917c68"),
-      ("NOTICE", "d64d250d1c3a5af25977651b5443aedb"),
-      ("resource1.txt", "bc9171172c4723589a247f99b838732d"),
-      ("resource2.txt", "9cd530447200979dbf9e117915cbcc74"),
-      ("resource_folder/resource_L1_1.txt", "51954bf10062451e683121e58d858417"),
-      ("resource_folder/resource_L1_2.txt", "b43991c0ef5d15710faf976e02cbb206"),
-      ("resource_folder/resource_L2/resource_L2_1.txt", "63165187f791a8dfff628ef8090e56ff"),
-      ("target_folder/relocated_file_1.txt", "bc9171172c4723589a247f99b838732d"),
-      ("target_folder/relocated_file_2.txt", "51954bf10062451e683121e58d858417"),
-      ("target_folder/relocated_file_3.txt", "6b0e05ae3d38b7db48ebdfc564366bce"),
-      ("resource3.txt", "aa2037b3d308bcb6a78a3d4fbf04b297"),
-      ("target_folder/relocated_file_4.txt", "aa2037b3d308bcb6a78a3d4fbf04b297")
-    )
-
-    // Check all resources can be found in the folder
-    for ((name, md5sum) <- expectedResources) {
-      val resourceFile = Paths.get(tempFolStr, name).toFile
-
-      assert(resourceFile.exists, s"Could not find $name")
-
-      val hash = TestHelper.computeHash(resourceFile.getPath)
-      assert(hash == md5sum, s"Calculated md5sum doesn't match the given md5sum for $name")
-    }
-  }
-
-  test("Check resources with unsupported format") {
-    // generate viash script
-    val testOutput = TestHelper.testMainException2[RuntimeException](
-      "build",
-      "-p", "docker",
-      "-o", tempFolStr,
-      configResourcesUnsupportedProtocolFile
-    )
-
-    assert(testOutput.exceptionText == "Unsupported scheme: ftp")
-  }
-  //</editor-fold>
-  //<editor-fold desc="Test docker options chown, port and workdir">
-  def docker_chown_get_owner(dockerId: String): String = {
-    val localConfig = configDockerOptionsChownFile
-    val localFunctionality = Config.read(localConfig, modifyFun = false).functionality
-    val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
-
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", dockerId,
-      "-o", tempFolStr,
-      "--setup", "build",
-      localConfig
-    )
-
-    assert(localExecutable.exists)
-    assert(localExecutable.canExecute)
-
-    // run the script
-    val output = Paths.get(tempFolStr, s"output_" + dockerId + ".txt").toFile
-
-    Exec.run(
-      Seq(
-        localExecutable.toString,
-        localExecutable.toString,
-        "--real_number", "10.5",
-        "--whole_number=10",
-        "-s", "a string with a few spaces",
-        "--output", output.getPath
-      )
-    )
-
-    assert(output.exists())
-
-    val owner = Files.getOwner(output.toPath)
-    owner.toString
-  }
-
-  def docker_chown_get_owner_two_outputs(dockerId: String): (String,String) = {
-    val localConfig = configDockerOptionsChownTwoOutputFile
-    val localFunctionality = Config.read(localConfig, modifyFun = false).functionality
-    val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
-
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", dockerId,
-      "-o", tempFolStr,
-      "--setup", "build",
-      localConfig
-    )
-
-    assert(localExecutable.exists)
-    assert(localExecutable.canExecute)
-
-    // run the script
-    val output = Paths.get(tempFolStr, "output_" + dockerId + ".txt").toFile
-    val output2 = Paths.get(tempFolStr, "output_" + dockerId +"_2.txt").toFile
-
-    val _ = Exec.run(
-      Seq(
-        localExecutable.toString,
-        localExecutable.toString,
-        "--real_number", "10.5",
-        "--whole_number=10",
-        "-s", "a string with a few spaces",
-        "--output", output.getPath,
-        "--output2", output2.getPath
-      )
-    )
-
-    assert(output.exists())
-    assert(output2.exists())
-
-    val owner = Files.getOwner(output.toPath)
-    val owner2 = Files.getOwner(output2.toPath)
-    (owner.toString, owner2.toString)
-  }
-
-  def docker_chown_get_owner_multiple_outputs(dockerId: String): (String,String,String) = {
-    val localConfig = configDockerOptionsChownMultipleOutputFile
-    val localFunctionality = Config.read(localConfig, modifyFun = false).functionality
-    val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
-
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", dockerId,
-      "-o", tempFolStr,
-      "--setup", "build",
-      localConfig
-    )
-
-    assert(localExecutable.exists)
-    assert(localExecutable.canExecute)
-
-    // run the script
-    val output = Paths.get(tempFolStr, "output_" + dockerId + ".txt").toFile
-    val output2 = Paths.get(tempFolStr, "output_" + dockerId +"_2.txt").toFile
-    val output3 = Paths.get(tempFolStr, "output_" + dockerId +"_3.txt").toFile
-
-    Exec.run(
-      Seq(
-        localExecutable.toString,
-        localExecutable.toString,
-        "--real_number", "10.5",
-        "--whole_number=10",
-        "-s", "a string with a few spaces",
-        "--output", output.getPath, output2.getPath, output3.getPath
-      )
-    )
-
-    assert(output.exists())
-    assert(output2.exists())
-    assert(output3.exists())
-
-    val owner = Files.getOwner(output.toPath)
-    val owner2 = Files.getOwner(output2.toPath)
-    val owner3 = Files.getOwner(output3.toPath)
-    (owner.toString, owner2.toString, owner3.toString)
-  }
-
-  test("Test default behaviour when chown is not specified", DockerTest) {
-    val owner = docker_chown_get_owner("chown_default")
-    assert(owner.nonEmpty)
-    assert(owner != "root")
-  }
-
-  test("Test default behaviour when chown is set to true", DockerTest) {
-    val owner = docker_chown_get_owner("chown_true")
-    assert(owner.nonEmpty)
-    assert(owner != "root")
-  }
-
-  test("Test default behaviour when chown is set to false", DockerTest) {
-    val owner = docker_chown_get_owner("chown_false")
-    assert(owner == "root")
-  }
-
-  test("Test default behaviour when chown is not specified with two output files", DockerTest) {
-    val owner = docker_chown_get_owner_two_outputs("two_chown_default")
-    assert(owner._1.nonEmpty)
-    assert(owner._2.nonEmpty)
-    assert(owner._1 != "root")
-    assert(owner._2 != "root")
-  }
-
-  test("Test default behaviour when chown is set to true with two output files", DockerTest) {
-    val owner = docker_chown_get_owner_two_outputs("two_chown_true")
-    assert(owner._1.nonEmpty)
-    assert(owner._2.nonEmpty)
-    assert(owner._1 != "root")
-    assert(owner._2 != "root")
-  }
-
-  test("Test default behaviour when chown is set to false with two output files", DockerTest) {
-    val owner = docker_chown_get_owner_two_outputs("two_chown_false")
-    assert(owner._1 == "root")
-    assert(owner._2 == "root")
-  }
-
-  test("Test default behaviour when chown is not specified with multiple output files", DockerTest) {
-    val owner = docker_chown_get_owner_multiple_outputs("multiple_chown_default")
-    assert(owner._1.nonEmpty)
-    assert(owner._2.nonEmpty)
-    assert(owner._3.nonEmpty)
-    assert(owner._1 != "root")
-    assert(owner._2 != "root")
-    assert(owner._3 != "root")
-  }
-
-  test("Test default behaviour when chown is set to true with multiple output files", DockerTest) {
-    val owner = docker_chown_get_owner_multiple_outputs("multiple_chown_true")
-    assert(owner._1.nonEmpty)
-    assert(owner._2.nonEmpty)
-    assert(owner._3.nonEmpty)
-    assert(owner._1 != "root")
-    assert(owner._2 != "root")
-    assert(owner._3 != "root")
-  }
-
-  test("Test default behaviour when chown is set to false with multiple output files", DockerTest) {
-    val owner = docker_chown_get_owner_multiple_outputs("multiple_chown_false")
-    assert(owner._1 == "root")
-    assert(owner._2 == "root")
-    assert(owner._3 == "root")
-  }
   //</editor-fold>
 
   def checkDockerImageExists(name: String): Boolean = {
