+++ src/test/scala/com/dataintuitive/viash/MainTestDockerSuite.scala
@@ -21,9 +21,6 @@ class MainTestDockerSuite extends FunSuite with BeforeAndAfterAll {
   private val configTextFile = getClass.getResource("/testbash/invalid_configs/config.txt").getPath
   private val configInvalidYamlFile = getClass.getResource("/testbash/invalid_configs/config_invalid_yaml.vsh.yaml").getPath
 
-  private val configResourcesCopyFile = getClass.getResource("/testbash/config_resource_test.vsh.yaml").getPath
-  private val configResourcesUnsupportedProtocolFile = getClass.getResource("/testbash/config_resource_unsupported_protocol.vsh.yaml").getPath
-
   // custom platform yamls
   private val customPlatformFile = getClass.getResource("/testbash/platform_custom.yaml").getPath
 
@@ -249,91 +246,6 @@ class MainTestDockerSuite extends FunSuite with BeforeAndAfterAll {
     assert(testText.contains("Running tests in temporary directory: "))
     assert(testText.contains("SUCCESS! All 2 out of 2 test scripts succeeded!"))
     assert(testText.contains("Cleaning up temporary directory"))
-  }
-  //</editor-fold>
-  //<editor-fold desc="Verify correct copying of resources">
-  test("Check resources are copied from and to the correct location", DockerTest) {
-
-    // copy some resources to /tmp/viash_tmp_resources/ so we can test absolute path resources
-    val tmpFolderResourceSourceFile = Paths.get(getClass.getResource("/testbash/resource3.txt").getFile)
-
-    val tmpFolderResourceDestinationFolder = Paths.get("/tmp/viash_tmp_resources/").toFile
-    val tmpFolderResourceDestinationFile = Paths.get(tmpFolderResourceDestinationFolder.getPath, "resource3.txt")
-
-    if (!tmpFolderResourceDestinationFolder.exists())
-      tmpFolderResourceDestinationFolder.mkdir()
-
-    Files.copy(tmpFolderResourceSourceFile, tmpFolderResourceDestinationFile, StandardCopyOption.REPLACE_EXISTING)
-
-    // generate viash script
-    val testText = TestHelper.testMain(
-      "test",
-      "-p", "docker",
-      "-k", "true",
-      configResourcesCopyFile
-    )
-
-    // basic checks to see if standard test/build was correct
-    assert(testText.contains("Running tests in temporary directory: "))
-    assert(testText.contains("WARNING! No tests found!"))
-    assert(!testText.contains("Cleaning up temporary directory"))
-
-    val FolderRegex = ".*Running tests in temporary directory: '([^']*)'.*".r
-
-    val tempPath = testText.replaceAll("\n", "") match {
-      case FolderRegex(path) => path
-      case _ => ""
-    }
-
-    // List all expected resources and their md5sum
-    val expectedResources = List(
-      //("check_bash_version.sh", "0c3c134d4ff0ea3a4a3b32e09fb7c100"),
-      ("code.sh", "efa9e1aa1c5f2a0b91f558ead5917c68"),
-      ("NOTICE", ".*"), // turn off checksum match for notice, as it might change.
-      ("resource1.txt", "bc9171172c4723589a247f99b838732d"),
-      ("resource2.txt", "9cd530447200979dbf9e117915cbcc74"),
-      ("resource_folder/resource_L1_1.txt", "51954bf10062451e683121e58d858417"),
-      ("resource_folder/resource_L1_2.txt", "b43991c0ef5d15710faf976e02cbb206"),
-      ("resource_folder/resource_L2/resource_L2_1.txt", "63165187f791a8dfff628ef8090e56ff"),
-      ("target_folder/relocated_file_1.txt", "bc9171172c4723589a247f99b838732d"),
-      ("target_folder/relocated_file_2.txt", "51954bf10062451e683121e58d858417"),
-      ("target_folder/relocated_file_3.txt", "6b0e05ae3d38b7db48ebdfc564366bce"),
-      ("resource3.txt", "aa2037b3d308bcb6a78a3d4fbf04b297"),
-      ("target_folder/relocated_file_4.txt", "aa2037b3d308bcb6a78a3d4fbf04b297")
-    )
-
-    // Check all resources can be found in the folder
-    for ((name, md5sum) <- expectedResources) {
-      val resourceFile = Paths.get(tempPath, "build_executable", name).toFile
-
-      assert(resourceFile.exists, s"Could not find $name")
-
-      val hash = TestHelper.computeHash(resourceFile.getPath)
-
-      assert(md5sum.r.findFirstMatchIn(hash).isDefined, s"Calculated md5sum doesn't match the given md5sum for $name")
-    }
-
-    Directory(tmpFolderResourceDestinationFolder).deleteRecursively()
-    checkTempDirAndRemove(testText, true)
-  }
-
-  test("Check resources with unsupported format", DockerTest) {
-    // generate viash script
-    val testOutput = TestHelper.testMainException2[RuntimeException](
-      "test",
-      "-p", "docker",
-      "-k", "true",
-      configResourcesUnsupportedProtocolFile
-    )
-
-    assert(testOutput.exceptionText == "Unsupported scheme: ftp")
-
-    // basic checks to see if standard test/build was correct
-    assert(testOutput.output.contains("Running tests in temporary directory: "))
-    assert(!testOutput.output.contains("WARNING! No tests found!"))
-    assert(!testOutput.output.contains("Cleaning up temporary directory"))
-
-    checkTempDirAndRemove(testOutput.output, true)
   }
   //</editor-fold>
